src/main/webapp/app/entities/def/route/def-routing-resolve.service.spec.ts
jest.mock('@angular/router');

import { TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ActivatedRouteSnapshot, Router } from '@angular/router';
import { of } from 'rxjs';

import { IDef, Def } from '../def.model';
import { DefService } from '../service/def.service';

import { DefRoutingResolveService } from './def-routing-resolve.service';

describe('Service Tests', () => {
  describe('Def routing resolve service', () => {
    let mockRouter: Router;
    let mockActivatedRouteSnapshot: ActivatedRouteSnapshot;
    let routingResolveService: DefRoutingResolveService;
    let service: DefService;
    let resultDef: IDef | undefined;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule],
        providers: [Router, ActivatedRouteSnapshot],
      });
      mockRouter = TestBed.inject(Router);
      mockActivatedRouteSnapshot = TestBed.inject(ActivatedRouteSnapshot);
      routingResolveService = TestBed.inject(DefRoutingResolveService);
      service = TestBed.inject(DefService);
      resultDef = undefined;
    });

    describe('resolve', () => {
      it('should return IDef returned by find', () => {
        // GIVEN
        service.find = jest.fn(id => of(new HttpResponse({ body: { id } })));
        mockActivatedRouteSnapshot.params = { id: 123 };

        // WHEN
        routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
          resultDef = result;
        });

        // THEN
        expect(service.find).toBeCalledWith(123);
        expect(resultDef).toEqual({ id: 123 });
      });

      it('should return new IDef if id is not provided', () => {
        // GIVEN
        service.find = jest.fn();
        mockActivatedRouteSnapshot.params = {};

        // WHEN
        routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
          resultDef = result;
        });

        // THEN
        expect(service.find).not.toBeCalled();
        expect(resultDef).toEqual(new Def());
      });

      it('should route to 404 page if data not found in server', () => {
        // GIVEN
        jest.spyOn(service, 'find').mockReturnValue(of(new HttpResponse({ body: null as unknown as Def })));
        mockActivatedRouteSnapshot.params = { id: 123 };

        // WHEN
        routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
          resultDef = result;
        });

        // THEN
        expect(service.find).toBeCalledWith(123);
        expect(resultDef).toEqual(undefined);
        expect(mockRouter.navigate).toHaveBeenCalledWith(['404']);
      });
    });
  });
});
